Path class to return path from Dijkstra, will be later used by Bellman-Ford. Saving/Loading graphs to text file (very basic, should be remade if it's supposed to be real feature)

+++ src/graph.cpp
@@ -6,12 +6,12 @@
 namespace Graphs
 {
 	
-	unsigned int Graph::size_edges(void) const
+	size_t Graph::size_edges(void) const
 	{
 		return this->_edges.size();
 	}
 
-	unsigned int Graph::size_nodes(void) const
+	size_t Graph::size_nodes(void) const
 	{
 		return this->_nodes.size();
 	}
